# Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mira API
  version: 1.0.0
  description: "API documentation for Mira: payments, webhooks, and admin partner management. Generated from repository routes."
servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: '3000'
paths:
...
  /api/payments/webhook/{gateway}:
    post:
      summary: Webhook receiver for payment gateway ( nomba)
      parameters:
        - name: gateway
          in: path
          required: true
          schema:
            type: string
            enum: [nomba]
      requestBody:
        required: true
        content:
          application/json: {}
      responses:
        '200':
          description: OK
        '401':
          description: Invalid signature
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...